Validate the route in the server and change decimal by double to save correctly lat and lng

+++ database/migrations/2019_05_21_205249_create_routes_table.php
@@ -16,12 +16,11 @@ class CreateRoutesTable extends Migration
         Schema::create('routes', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->decimal('lat_start');
-            $table->decimal('lng_start');
+            $table->double('lat_start');
+            $table->double('lng_start');
             $table->string('formatted_address_start');
-
-            $table->decimal('lat_end');
-            $table->decimal('lng_end');
+            $table->double('lat_end');
+            $table->double('lng_end');
             $table->string('formatted_address_end');
 
             $table->unsignedBigInteger('user_id')->index();
